reorganized module structure
